+++ aztec/src/main/kotlin/org/wordpress/aztec/formatting/BlockFormatter.kt
@@ -42,7 +42,7 @@ class BlockFormatter(editor: AztecText, val listStyle: ListStyle, val quoteStyle
     }
 
     fun toggleQuote() {
-        if (!containQuote()) {
+        if (!containsQuote()) {
             applyBlockStyle(AztecTextFormat.FORMAT_QUOTE)
         } else {
             removeBlockStyle(AztecTextFormat.FORMAT_QUOTE)
@@ -134,7 +134,7 @@ class BlockFormatter(editor: AztecText, val listStyle: ListStyle, val quoteStyle
         val boundsOfSelectedText = if (ignoreLineBounds) {
             IntRange(start, end)
         } else {
-            getSelectedTextBounds(editableText, start, end)
+            getBoundsOfText(editableText, start, end)
         }
 
         var startOfBounds = boundsOfSelectedText.start
@@ -296,34 +296,59 @@ class BlockFormatter(editor: AztecText, val listStyle: ListStyle, val quoteStyle
         }
     }
 
-    fun getSelectedTextBounds(editable: Editable, selectionStart: Int, selectionEnd: Int): IntRange {
-        val startOfLine: Int
-        val endOfLine: Int
+    /**
+     * Returns paragraph bounds (\n) to the left and to the right of selection.
+     */
+    fun getBoundsOfText(editable: Editable, selectionStart: Int, selectionEnd: Int): IntRange {
+        val startOfBlock: Int
+        val endOfBlock: Int
+
+        val selectionStartIsOnTheNewLine = selectionStart != selectionEnd && selectionStart > 0
+                && selectionStart < editableText.length
+                && editable[selectionStart] == '\n'
+
+        val selectionStartIsBetweenNewlines = selectionStartIsOnTheNewLine
+                && selectionStart > 0
+                && selectionStart < editableText.length
+                && editable[selectionStart - 1] == '\n'
+
+        val isTrailingNewlineAtTheEndOfSelection = selectionStart != selectionEnd
+                && selectionEnd > 0
+                && editableText.length > selectionEnd
+                && editableText[selectionEnd] != Constants.END_OF_BUFFER_MARKER
+                && editableText[selectionEnd] != '\n'
+                && editableText[selectionEnd - 1] == '\n'
 
         val indexOfFirstLineBreak: Int
-        val indexOfLastLineBreak = editable.indexOf("\n", selectionEnd)
+        var indexOfLastLineBreak = editable.indexOf("\n", selectionEnd)
 
-        if (indexOfLastLineBreak > 0) {
-            val characterBeforeLastLineBreak = editable[indexOfLastLineBreak - 1]
-            if (characterBeforeLastLineBreak != '\n') {
-                indexOfFirstLineBreak = editable.lastIndexOf("\n", selectionStart - 1) + 1
-            } else {
-                indexOfFirstLineBreak = editable.lastIndexOf("\n", selectionStart)
+
+        if (selectionStartIsBetweenNewlines) {
+            indexOfFirstLineBreak = selectionStart
+        } else if (selectionStartIsOnTheNewLine) {
+            val isSingleCharacterLine = (selectionStart > 1 && editableText[selectionStart - 1] != '\n' && editableText[selectionStart - 2] == '\n') || selectionStart == 1
+            indexOfFirstLineBreak = editable.lastIndexOf("\n", selectionStart - if (isSingleCharacterLine) 0 else 1) - if (isSingleCharacterLine) 1 else 0
+
+            if (isTrailingNewlineAtTheEndOfSelection) {
+                indexOfLastLineBreak = editable.indexOf("\n", selectionEnd - 1)
+            }
+        } else if (isTrailingNewlineAtTheEndOfSelection) {
+            indexOfFirstLineBreak = editable.lastIndexOf("\n", selectionStart - 1) + 1
+            indexOfLastLineBreak = editable.indexOf("\n", selectionEnd - 1)
+        } else if (indexOfLastLineBreak > 0) {
+            indexOfFirstLineBreak = editable.lastIndexOf("\n", selectionStart - 1) + 1
+        } else if (indexOfLastLineBreak == -1) {
+            indexOfFirstLineBreak = if (selectionStart == 0) 0 else {
+                editable.lastIndexOf("\n", selectionStart) + 1
             }
         } else {
-            if (indexOfLastLineBreak == -1) {
-                indexOfFirstLineBreak = if (selectionStart == 0) 0 else {
-                    editable.lastIndexOf("\n", selectionStart) + 1
-                }
-            } else {
-                indexOfFirstLineBreak = editable.lastIndexOf("\n", selectionStart)
-            }
+            indexOfFirstLineBreak = editable.lastIndexOf("\n", selectionStart)
         }
 
-        startOfLine = if (indexOfFirstLineBreak != -1) indexOfFirstLineBreak else 0
-        endOfLine = if (indexOfLastLineBreak != -1) (indexOfLastLineBreak + 1) else editable.length
+        startOfBlock = if (indexOfFirstLineBreak != -1) indexOfFirstLineBreak else 0
+        endOfBlock = if (indexOfLastLineBreak != -1) (indexOfLastLineBreak + 1) else editable.length
 
-        return IntRange(startOfLine, endOfLine)
+        return IntRange(startOfBlock, endOfBlock)
     }
 
     fun applyBlockStyle(blockElementType: ITextFormat, start: Int = selectionStart, end: Int = selectionEnd) {
@@ -333,42 +358,27 @@ class BlockFormatter(editor: AztecText, val listStyle: ListStyle, val quoteStyle
         }
 
         if (start != end) {
-            val nestingLevel = IAztecNestable.getNestingLevelAt(editableText, start)
+            val nestingLevelAtTheStartOfSelection = IAztecNestable.getNestingLevelAt(editableText, start)
+            val nestingLevelAtTheEndOfSelection = IAztecNestable.getNestingLevelAt(editableText, end)
 
-            if (IAztecNestable.getNestingLevelAt(editableText, end) != nestingLevel) {
-                // TODO: styling across multiple nesting levels not support yet
-                return
-            }
-
-            val indexOfFirstLineBreak = editableText.indexOf("\n", end)
-
-            val endOfBlock = if (indexOfFirstLineBreak != -1) indexOfFirstLineBreak else editableText.length
-            val startOfBlock = editableText.lastIndexOf("\n", start)
-
-            val selectedLines = editableText.subSequence(startOfBlock + 1..endOfBlock - 1) as Editable
-
-            var numberOfLinesWithSpanApplied = 0
-            var numberOfLines = 0
-
-            val lines = TextUtils.split(selectedLines.toString(), "\n")
-
-            for (i in lines.indices) {
-                numberOfLines++
-                if (containsList(blockElementType, i, selectedLines, nestingLevel)) {
-                    numberOfLinesWithSpanApplied++
+            // TODO: styling across multiple nesting levels not support yet
+            if (nestingLevelAtTheStartOfSelection != nestingLevelAtTheEndOfSelection) {
+                if (nestingLevelAtTheStartOfSelection == 0 && nestingLevelAtTheEndOfSelection == 1) {
+                    // 0/1 is ok!
+                } else {
+                    return
                 }
             }
 
-            if (numberOfLines == numberOfLinesWithSpanApplied) {
-                removeBlockStyle(blockElementType)
-            } else {
-                //if block starts with newline do not move index to the right
-                val startOfBlockModifier = if (startOfBlock >= 0 && editableText[startOfBlock] == '\n') 0 else 1
-                applyBlock(makeBlockSpan(blockElementType, nestingLevel), startOfBlock + startOfBlockModifier,
-                        (if (endOfBlock == editableText.length) endOfBlock else endOfBlock + 1))
-            }
+            val boundsOfSelectedText = getBoundsOfText(editableText, start, end)
+
+            val startOfBlock = boundsOfSelectedText.start
+            val endOfBlock = boundsOfSelectedText.endInclusive
+
+            applyBlock(makeBlockSpan(blockElementType, nestingLevelAtTheStartOfSelection), startOfBlock,
+                    (if (endOfBlock == editableText.length) endOfBlock else endOfBlock))
         } else {
-            val boundsOfSelectedText = getSelectedTextBounds(editableText, start, end)
+            val boundsOfSelectedText = getBoundsOfText(editableText, start, end)
 
             val startOfLine = boundsOfSelectedText.start
             val endOfLine = boundsOfSelectedText.endInclusive
@@ -517,10 +527,10 @@ class BlockFormatter(editor: AztecText, val listStyle: ListStyle, val quoteStyle
 
         if (list.isEmpty()) return false
 
-        return list.any { containsList(textFormat, it, editableText, nestingLevel) }
+        return list.any { containsBlockElement(textFormat, it, editableText, nestingLevel) }
     }
 
-    fun containsList(textFormat: ITextFormat, index: Int, text: Editable, nestingLevel: Int): Boolean {
+    fun containsBlockElement(textFormat: ITextFormat, index: Int, text: Editable, nestingLevel: Int): Boolean {
         val lines = TextUtils.split(text.toString(), "\n")
         if (index < 0 || index >= lines.size) {
             return false
@@ -537,53 +547,26 @@ class BlockFormatter(editor: AztecText, val listStyle: ListStyle, val quoteStyle
         return spans.isNotEmpty()
     }
 
-    fun containQuote(selStart: Int = selectionStart, selEnd: Int = selectionEnd): Boolean {
-        val lines = TextUtils.split(editableText.toString(), "\n")
-        val list = ArrayList<Int>()
+    fun containsQuote(selStart: Int = selectionStart, selEnd: Int = selectionEnd): Boolean {
+        if (selStart < 0 || selEnd < 0) return false
 
-        for (i in lines.indices) {
-            val lineStart = (0..i - 1).sumBy { lines[it].length + 1 }
-            val lineEnd = lineStart + lines[i].length
+        return editableText.getSpans(selStart, selEnd, AztecQuoteSpan::class.java)
+                .any {
+                    val spanStart = editableText.getSpanStart(it)
+                    val spanEnd = editableText.getSpanEnd(it)
 
-            if (lineStart >= lineEnd) {
-                continue
-            }
+                    if (selStart == selEnd) {
+                        if (editableText.length == selStart) {
+                            selStart in spanStart..spanEnd
+                        } else {
+                            (spanEnd != selStart) && selStart in spanStart..spanEnd
+                        }
 
-            /**
-             * lineStart  >= selStart && selEnd   >= lineEnd // single line, current entirely selected OR
-             *                                                  multiple lines (before and/or after), current entirely selected
-             * lineStart  <= selEnd   && selEnd   <= lineEnd // single line, current partially or entirely selected OR
-             *                                                  multiple lines (after), current partially or entirely selected
-             * lineStart  <= selStart && selStart <= lineEnd // single line, current partially or entirely selected OR
-             *                                                  multiple lines (before), current partially or entirely selected
-             */
-            if ((lineStart >= selStart && selEnd >= lineEnd)
-                    || (lineStart <= selEnd && selEnd <= lineEnd)
-                    || (lineStart <= selStart && selStart <= lineEnd)) {
-                list.add(i)
-            }
-        }
-
-        if (list.isEmpty()) return false
-
-        return list.any { containQuote(it) }
-    }
-
-    fun containQuote(index: Int): Boolean {
-        val lines = TextUtils.split(editableText.toString(), "\n")
-        if (index < 0 || index >= lines.size) {
-            return false
-        }
-
-        val start = (0..index - 1).sumBy { lines[it].length + 1 }
-        val end = start + lines[index].length
-
-        if (start >= end) {
-            return false
-        }
-
-        val spans = editableText.getSpans(start, end, AztecQuoteSpan::class.java)
-        return spans.isNotEmpty()
+                    } else {
+                        (selStart in spanStart..spanEnd || selEnd in spanStart..spanEnd) ||
+                                (spanStart in selStart..selEnd || spanEnd in spanStart..spanEnd)
+                    }
+                }
     }
 
     fun containsHeading(textFormat: ITextFormat, selStart: Int = selectionStart, selEnd: Int = selectionEnd): Boolean {